+++ src/app/beer_garden/api/http/handlers/v1/topic.py
@@ -63,7 +63,7 @@ class TopicAPI(BaseHandler):
           - Topics
         """
 
-        await self.client(Operation(operation_type="TOPIC_DELETE", args=[topic_id]))
+        await self.client(Operation(operation_type="TOPIC_DELETE", kwargs={"topic_id": topic_id}))
 
         self.set_status(204)
 
@@ -138,6 +138,136 @@ class TopicAPI(BaseHandler):
         self.set_header("Content-Type", "application/json; charset=UTF-8")
         self.write(response)
 
+class TopicNameAPI(BaseHandler):
+    parser = SchemaParser()
+
+    async def get(self, topic_name):
+        """
+        ---
+        summary: Get a topic_name by id
+        parameters:
+          - name: topic_name
+            in: path
+            required: true
+            description: The name of the Topic
+            type: string
+        responses:
+          200:
+            description: List of topic states
+            schema:
+              $ref: '#/definitions/Topic'
+          404:
+            $ref: '#/definitions/404Error'
+          50x:
+            $ref: '#/definitions/50xError'
+        tags:
+          - Topics
+        """
+
+        response = await self.client(
+            Operation(operation_type="TOPIC_READ", kwargs={"topic_name": topic_name})
+        )
+
+        self.set_header("Content-Type", "application/json; charset=UTF-8")
+        self.write(response)
+
+    async def delete(self, topic_name):
+        """
+        ---
+        summary: Delete a topic
+        parameters:
+          - name: topic_name
+            in: path
+            required: true
+            description: The name of the topic
+            type: string
+        responses:
+          200:
+            description: List of topic states
+            schema:
+              $ref: '#/definitions/Topic'
+          404:
+            $ref: '#/definitions/404Error'
+          50x:
+            $ref: '#/definitions/50xError'
+        tags:
+          - Topics
+        """
+
+        await self.client(Operation(operation_type="TOPIC_DELETE", kwargs={"topic_name": topic_name}))
+
+        self.set_status(204)
+
+    async def patch(self, topic_name):
+        """
+        ---
+        summary: Partially update a Topic
+        description: |
+          The body of the request needs to contain a set of instructions detailing the
+          updates to apply. Currently the only operations are:
+
+          * add
+          * remove
+
+          ```JSON
+          [
+            { "operation": "add", "value": {subscriber} }
+            { "operation": "remove", "value": {subscriber} }
+          ]
+          ```
+        parameters:
+          - name: topic_name
+            in: path
+            required: true
+            description: The name of the Topic
+            type: string
+          - name: patch
+            in: body
+            required: true
+            description: Instructions for how to update the Topic
+            schema:
+              $ref: '#/definitions/Patch'
+        responses:
+          200:
+            description: Topic with the given name
+            schema:
+              $ref: '#/definitions/Topic'
+          400:
+            $ref: '#/definitions/400Error'
+          404:
+            $ref: '#/definitions/404Error'
+          50x:
+            $ref: '#/definitions/50xError'
+        tags:
+          - Topics
+        """
+        patch = SchemaParser.parse_patch(self.request.decoded_body, from_string=True)
+
+        for op in patch:
+            operation = op.operation.lower()
+            subscriber = BrewtilsSubscriber(**op.value)
+
+            if operation == "add":
+                response = await self.client(
+                    Operation(
+                        operation_type="TOPIC_ADD_SUBSCRIBER",
+                        kwargs={"topic_name": topic_name, "subscriber": subscriber},
+                    )
+                )
+
+            elif operation == "remove":
+                response = await self.client(
+                    Operation(
+                        operation_type="TOPIC_REMOVE_SUBSCRIBER",
+                        kwargs={"topic_name": topic_name, "subscriber": subscriber},
+                    )
+                )
+
+            else:
+                raise ModelValidationError(f"Unsupported operation '{op.operation}'")
+
+        self.set_header("Content-Type", "application/json; charset=UTF-8")
+        self.write(response)
 
 class TopicListAPI(BaseHandler):
     parser = SchemaParser()